requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
databases[postgresql]==0.5.7
SQLAlchemy==2.0.19
asyncpg==0.27.0
httpx==0.24.1
python-dotenv==1.0.0
dateutil==2.8.2
